updated discovery code
